+++ CHANGELOG.rst
@@ -2,6 +2,12 @@ v0.3.0.0
 --------
 
 * Code refactor to follow PEP8
+* All new Markov Chain module using numpy backend
+    * Function definitions largely the same for common functions
+* Functions renamed:
+    * Table.expected_value -> Table.ev
+    * Table.variance -> Table.var
+* Unit Tests!
 
 v0.2.9.0 (2017-03-19)
 ---------------------